+++ sdk/src/client/api/block_builder/input_selection/core/remainder.rs
@@ -15,10 +15,7 @@ use crate::{
     client::api::RemainderData,
     types::block::{
         address::{Address, Ed25519Address},
-        output::{
-            unlock_condition::{AddressUnlockCondition, UnlockCondition},
-            BasicOutputBuilder, NativeTokensBuilder, Output,
-        },
+        output::{unlock_condition::AddressUnlockCondition, BasicOutputBuilder, NativeTokensBuilder, Output},
     },
 };
 
@@ -65,8 +62,8 @@ impl InputSelection {
 
         let mut remainder_builder =
             BasicOutputBuilder::new_with_minimum_storage_deposit(self.protocol_parameters.rent_structure().clone())?
-                .add_unlock_condition(UnlockCondition::Address(AddressUnlockCondition::new(Address::from(
-                    Ed25519Address::from([0; 32]),
+                .add_unlock_condition(AddressUnlockCondition::new(Address::from(Ed25519Address::from(
+                    [0; 32],
                 ))));
 
         if let Some(native_tokens) = native_tokens_diff {
@@ -94,7 +91,7 @@ impl InputSelection {
             if amount > output_sdr_amount {
                 let diff = amount - output_sdr_amount;
                 let srd_output = BasicOutputBuilder::new_with_amount(diff)?
-                    .with_unlock_conditions([UnlockCondition::Address(AddressUnlockCondition::new(address))])
+                    .with_unlock_conditions([AddressUnlockCondition::new(address)])
                     .finish_output(self.protocol_parameters.token_supply())?;
 
                 // TODO verify_storage_deposit ?
@@ -132,8 +129,7 @@ impl InputSelection {
         let diff = inputs_sum - outputs_sum;
         let mut remainder_builder = BasicOutputBuilder::new_with_amount(diff)?;
 
-        remainder_builder = remainder_builder
-            .add_unlock_condition(UnlockCondition::Address(AddressUnlockCondition::new(remainder_address)));
+        remainder_builder = remainder_builder.add_unlock_condition(AddressUnlockCondition::new(remainder_address));
 
         if let Some(native_tokens) = native_tokens_diff {
             log::debug!("Adding {native_tokens:?} to remainder output for {remainder_address:?}");