+++ sdk/src/wallet/account/mod.rs
@@ -198,7 +198,7 @@ fn serialize() {
     let output = Output::Basic(
         BasicOutput::build_with_amount(amount)
             .unwrap()
-            .add_unlock_condition(AddressUnlockCondition::new(address).into())
+            .add_unlock_condition(AddressUnlockCondition::new(address))
             .finish(protocol_parameters.token_supply())
             .unwrap(),
     );