+++ sdk/src/wallet/account/operations/transaction/high_level/send_micro_transaction.rs
@@ -9,7 +9,7 @@ use crate::{
         address::Address,
         output::{
             unlock_condition::{
-                AddressUnlockCondition, ExpirationUnlockCondition, StorageDepositReturnUnlockCondition, UnlockCondition,
+                AddressUnlockCondition, ExpirationUnlockCondition, StorageDepositReturnUnlockCondition,
             },
             BasicOutputBuilder,
         },
@@ -115,19 +115,19 @@ impl AccountHandle {
             outputs.push(
                 // Add address_and_amount.amount+storage_deposit_amount, so receiver can get address_and_amount.amount
                 BasicOutputBuilder::new_with_amount(address_with_amount.amount + storage_deposit_amount)?
-                    .add_unlock_condition(UnlockCondition::Address(AddressUnlockCondition::new(address)))
-                    .add_unlock_condition(UnlockCondition::StorageDepositReturn(
+                    .add_unlock_condition(AddressUnlockCondition::new(address))
+                    .add_unlock_condition(
                         // We send the storage_deposit_amount back to the sender, so only the additional amount is sent
                         StorageDepositReturnUnlockCondition::new(
                             return_address.address.inner,
                             storage_deposit_amount,
                             token_supply,
                         )?,
-                    ))
-                    .add_unlock_condition(UnlockCondition::Expiration(ExpirationUnlockCondition::new(
+                    )
+                    .add_unlock_condition(ExpirationUnlockCondition::new(
                         return_address.address.inner,
                         expiration_time,
-                    )?))
+                    )?)
                     .finish_output(token_supply)?,
             )
         }