+++ sdk/src/wallet/account/operations/transaction/mod.rs
@@ -40,9 +40,9 @@ impl AccountHandle {
     /// ```ignore
     /// let outputs = vec![
     ///    BasicOutputBuilder::new_with_amount(1_000_000)?
-    ///    .add_unlock_condition(UnlockCondition::Address(AddressUnlockCondition::new(
+    ///    .add_unlock_condition(AddressUnlockCondition::new(
     ///        Address::try_from_bech32("rms1qpszqzadsym6wpppd6z037dvlejmjuke7s24hm95s9fg9vpua7vluaw60xu")?,
-    ///    )))
+    ///    ))
     ///    .finish_output(account_handle.client.get_token_supply().await?;)?,
     /// ];
     /// let tx = account_handle