+++ sdk/CHANGELOG.md
@@ -19,6 +19,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
+## 0.2.0 - 2023-XX-XX
+
+### Changed
+
+- Builder methods `add_unlock_condition`, `replace_unlock_condition`, `with_unlock_conditions` are now generic;
+- Builder methods `add_feature`, `replace_feature`, `with_features` are now generic;
+- Builder methods `add_immutable_feature`, `replace_immutable_feature`, `with_immutable_features` are now generic;
+
 ## 0.1.0 - 2023-04-03
 
 First release of the `iota-sdk` crate.